build: Turn version.py into idiomatic Python
While it's possible to write C code in Python, it's better to actually
write Python code in Python.

Use regular expressions, instead of counting characters, to allow a
little bit more leeway when editing the cairo-version.h header file.

Use a context manager to handle the lifetime of a file object.

Use f-strings instead of the obsolete format() method.