meson: always skip IPC_RMID_DEFERRED_RELEASE check when cross-compiling
../../br-test-pkg/bootlin-armv5-uclibc/build/cairo-1.17.4/meson.build:279:13:
ERROR: Can not run test applications in this cross environment.

Commit 1bec56ea8a931e1ae1c74cc740134497ec365267 added support to define
ipc_rmid_deferred_release in a cross-compile config, but still kept
the default to auto, which anyhow results in an error when cross-compiling.

There is only one usage of the ipc_rmid_deferred_release compile declarative
which was originally added in this commit: 5041b462d084de8552336275914d30c23bf5dd35.

If ipc_rmid_deferred_release is set to FALSE, an additional XSync is performed.
This doesn't sound very harmful, so that is why this commit defaults to FALSE
and thus avoids any cross-compile errors.

Signed-off-by: Thomas Devoogdt <thomas@devoogdt.com>